gptactions/requirements.txt
Flask==2.1.2
gspread==6.1.0
oauth2client==4.1.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.38.0
Flask-Cors==3.0.10
